10.4: Massive commit; brings 10.4 tree up-to-date with 10.5 features (to the extent possible).
git-svn-id: http://macfuse.googlecode.com/svn/trunk@1047 febdc8c4-1326-0410-abd6-27cb15717650

+++ core/10.4/fusefs/fuse_ipc.c
@@ -41,7 +41,7 @@ static fuse_handler_t  fuse_standard_handler;
 void
 fiov_init(struct fuse_iov *fiov, size_t size)
 {
-    uint32_t msize = FU_AT_LEAST(size);
+    size_t msize = FU_AT_LEAST(size);
 
     fiov->len = 0;
 
@@ -335,7 +335,7 @@ int
 fticket_aw_pull_uio(struct fuse_ticket *ftick, uio_t uio)
 {
     int err = 0;
-    size_t len = uio_resid(uio);
+    size_t len = (size_t)uio_resid(uio);
 
     if (len) {
         switch (ftick->tk_aw_type) {
@@ -346,7 +346,7 @@ fticket_aw_pull_uio(struct fuse_ticket *ftick, uio_t uio)
                 IOLog("MacFUSE: failed to pull uio (error=%d)\n", err);
                 break;
             }
-            err = uiomove(fticket_resp(ftick)->base, len, uio);
+            err = uiomove(fticket_resp(ftick)->base, (int)len, uio);
             if (err) {
                 IOLog("MacFUSE: FT_A_FIOV error is %d (%p, %ld, %p)\n",
                       err, fticket_resp(ftick)->base, len, uio);
@@ -355,7 +355,7 @@ fticket_aw_pull_uio(struct fuse_ticket *ftick, uio_t uio)
 
         case FT_A_BUF:
             ftick->tk_aw_bufsize = len;
-            err = uiomove(ftick->tk_aw_bufdata, len, uio);
+            err = uiomove(ftick->tk_aw_bufdata, (int)len, uio);
             if (err) {
                 IOLog("MacFUSE: FT_A_BUF error is %d (%p, %ld, %p)\n",
                       err, ftick->tk_aw_bufdata, len, uio);
@@ -379,7 +379,7 @@ fticket_pull(struct fuse_ticket *ftick, uio_t uio)
         return 0;
     }
 
-    err = fuse_body_audit(ftick, uio_resid(uio));
+    err = fuse_body_audit(ftick, (size_t)uio_resid(uio));
     if (!err) {
         err = fticket_aw_pull_uio(ftick, uio);
     }
@@ -446,8 +446,8 @@ fdata_destroy(struct fuse_data *data)
     lck_mtx_free(data->aw_mtx, fuse_lock_group);
     data->aw_mtx = NULL;
 
-    lck_mtx_free(data->ticket_mtx, fuse_lock_group); /* XXX: can do? */
-    data->ticket_mtx = NULL;                         /* XXX: can do? */
+    lck_mtx_free(data->ticket_mtx, fuse_lock_group);
+    data->ticket_mtx = NULL;
 
 #if M_MACFUSE_EXPLICIT_RENAME_LOCK
     lck_rw_free(data->rename_lock, fuse_lock_group);
@@ -900,7 +900,7 @@ fuse_setup_ihead(struct fuse_in_header *ihead,
                  size_t                 blen,
                  vfs_context_t          context)
 {
-    ihead->len = sizeof(*ihead) + blen;
+    ihead->len = (uint32_t)(sizeof(*ihead) + blen);
     ihead->unique = ftick->tk_unique;
     ihead->nodeid = nid;
     ihead->opcode = op;
@@ -910,7 +910,7 @@ fuse_setup_ihead(struct fuse_in_header *ihead,
         ihead->uid = vfs_context_ucred(context)->cr_uid;
         ihead->gid = vfs_context_ucred(context)->cr_gid;
     } else {
-        /* XXX: more thought */
+        /* XXX: could use more thought */
         ihead->pid = proc_pid((proc_t)current_proc());
         ihead->uid = kauth_cred_getuid(kauth_cred_get());
         ihead->gid = kauth_cred_getgid(kauth_cred_get());
